Move the AssetRegister class into Boundary namespace
