+++ asset-register-api/Startup.cs
@@ -1,4 +1,4 @@
-﻿using System;
+﻿using hear_api.Boundary;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
