+++ asset-register-tests/HomesEngland/AcceptanceTest/GetAssetAcceptanceTest.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using asset_register_api.HomesEngland.Domain;
 using asset_register_api.Boundary.UseCase;
-using hear_api;
+using hear_api.Boundary;
 using NUnit.Framework;
 
 namespace asset_register_tests.HomesEngland.AcceptanceTest
