+++ asset-register-tests/HomesEngland/AcceptanceTest/SearchAssetsAcceptanceTest.cs
@@ -1,12 +1,9 @@
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using asset_register_api.HomesEngland.Domain;
-using asset_register_api.HomesEngland.UseCase;
 using asset_register_api.Boundary;
 using asset_register_api.Boundary.UseCase;
-using hear_api;
-using hear_api.HomesEngland.Gateway;
+using hear_api.Boundary;
 using NUnit.Framework;
 
 namespace asset_register_tests.HomesEngland.AcceptanceTest
